🎨 STRUCTURE: improved trpc structure
