+++ src/pages/api/trpc/[trpc].ts
@@ -1,6 +1,6 @@
 import { createNextApiHandler } from "@trpc/server/adapters/next";
-import { appRouter } from "@/server/trpc/routers/_app";
-import { createContext } from "@/server/trpc/context";
+import { appRouter } from "@/server/routers/_app";
+import { createContext } from "@/server/context";
 
 export default createNextApiHandler({
   router: appRouter,
