--- src/server/common/get-auth-session.ts
@@ -1,25 +1,0 @@
-import type {
-  NextApiRequest,
-  NextApiResponse,
-  GetServerSidePropsContext,
-} from "next";
-import { unstable_getServerSession } from "next-auth";
-
-import { authOptions } from "../../pages/api/auth/[...nextauth]";
-
-/**
- * Wrapper for unstable_getServerSession https://next-auth.js.org/configuration/nextjs
- * See example usage in trpc createContext or the restricted API route
- */
-const getServerAuthSession = async (ctx: {
-  req: GetServerSidePropsContext["req"];
-  res: GetServerSidePropsContext["res"];
-}) => {
-  return await unstable_getServerSession(
-    ctx.req,
-    ctx.res,
-    authOptions(ctx.req as NextApiRequest, ctx.res as NextApiResponse),
-  );
-};
-
-export default getServerAuthSession;
